improve modularity of network modules
Keep track of parameters and gradients in Module superclass and add
generic gradient related functions to the superclass. Furthermore, let
the `backwards` functions accumulate gradients, which can then be
applied to the parameters using said generic functions, and let the
`forwards` functions store activation values needed by `backwards` as
class variables.
As a result of these changes sequential modules can be pieced together
in a modular fashion, (mostly) only requiring the module output/gradient
to be passed on to the next/previous model `forward`/`backward`.